+++ server/buildoutls/types.py
@@ -37,3 +37,17 @@ class OpenPypiPageCommandParams(TypedDict):
 class UpdateMD5SumCommandParams(TypedDict):
   document_uri: str
   section_name: str
+
+
+# https://microsoft.github.io/language-server-protocol/specifications/lsp/3.18/specification/#textDocument_semanticTokens
+
+SEMANTIC_TOKEN_TYPES = [
+  "comment",
+  "string",
+  "number",
+  "keyword",
+  "class",
+  "function",
+  "variable",
+  "type",
+]